+++ TWLight/users/helpers/bundle_authorizations.py
@@ -3,7 +3,7 @@ from TWLight.users.models import Authorization
 
 
 # We have to put this in a new bundle_authorizations.py helper
-# rather than authorizations.py to avoid circular imports
+# rather than validation.py to avoid circular imports
 def get_all_bundle_authorizations():
     """
     Returns all Bundle authorizations, both active
