+++ TWLight/users/tests.py
@@ -23,7 +23,7 @@ from TWLight.resources.tests import EditorCraftRoom
 
 from . import views
 from .oauth import OAuthBackend
-from .helpers.authorizations import validate_partners
+from .helpers.validation import validate_partners
 from .helpers.bundle_authorizations import get_all_bundle_authorizations
 from .helpers.wiki_list import WIKIS, LANGUAGE_CODES
 from .factories import EditorFactory, UserFactory
